More cleanup Migrated to CMake Finished rewriting matrix library Cleaned up directories

+++ ssm/detail/type-traits.hpp
@@ -15,7 +15,7 @@ namespace simd
 {
 // value is true if the specified vector type can be accelerated using SIMD,
 // for example is_simd<float, 4> or is_simd<double, 2>
-template<typename T, size_t N>
+template<typename T, int N>
 struct is_simd { static constexpr bool value = false; };
 }
 }
